pyproject.toml
[project]
name = "orchestrator-agent"
version = "0.1.0"
description = "Standalone WFO search agent for Kagent deployment"
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "orchestrator-core[search]",
    "pydantic-ai-slim>=1.56.0",
    "ag-ui-protocol>=0.1.10",
    "fasta2a>=0.1.0",
    "mcp>=1.9.0",
    "uvicorn[standard]>=0.30.0",
    "structlog>=24.0.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.9",
    "pre-commit>=4.2.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "httpx>=0.27",
    "ruff>=0.12.0",
]
demo = [
    "crewai[a2a]",
]

[build-system]
requires = ["flit_core>=3.2,<4"]
build-backend = "flit_core.buildapi"

[tool.flit.module]
name = "orchestrator_agent"

[tool.black]
line-length = 120
target-version = ["py311"]
extend-exclude = "demos"

[tool.ruff]
exclude = [
    ".git",
    ".cache",
    ".eggs",
    "__pycache__",
    "demos",
    "dist",
    "venv",
    ".venv",
]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
ignore = [
    "C417",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D202",
    "E501",
    "N806",
    "B905",
    "N805",
    "B904",
    "N803",
    "N801",
    "N815",
    "N802",
]
select = ["B", "C", "D", "E", "F", "I", "N", "RET", "S", "T", "W"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "B033", "N816", "N802", "T201"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-third-party = ["pydantic"]
known-first-party = ["orchestrator_agent"]

[tool.pytest.ini_options]
norecursedirs = ".git build dist *.egg __pycache__ .cache .venv"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
exclude = "(venv|.venv|demos)"
ignore_missing_imports = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
namespace_packages = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_no_return = true
warn_unreachable = true
implicit_reexport = false
strict_equality = true
show_error_codes = true
show_column_numbers = true
plugins = ["pydantic.mypy"]
disable_error_code = "annotation-unchecked"

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_calls = false
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.uv]
index-url = "https://pypi.org/simple"
